requirements.txt
Flask==3.0.0
Flask-Session==0.5.0
requests==2.31.0
Werkzeug==3.0.1
pdfplumber==0.10.3
tabula-py==2.9.0
pandas==2.1.4
camelot-py[cv]==0.11.0
pytesseract==0.3.10
opencv-python==4.8.1.78
openpyxl==3.1.2
python-docx==1.1.0
PyMuPDF==1.23.8
pdf2image==1.16.3
reportlab==4.0.7
WeasyPrint==60.1
pdfkit==1.0.0
xlsxwriter==3.1.9
Pillow==10.1.0
python-pptx==0.6.23
numpy==1.26.2
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
